Test fixes and code cleanup
