+++ internal/config/cue.go
@@ -14,7 +14,7 @@ var schemaFile string
 // NewCue loads a cue configuration from filename and returns a Config struct.
 // The configuration is validated against the schema embedded in the binary.
 // TODO: For now, this config extends NewViper to extract default values. This should be in schema.cue.
-func NewCue(filename string) (*HostConfig, error) {
+func NewCue(filename string, base *ExperimentConfig) (*ExperimentConfig, error) {
 	ctx := cuecontext.New()
 	schema := ctx.CompileString(schemaFile).LookupPath(cue.ParsePath("config"))
 	if schema.Err() != nil {
@@ -32,12 +32,13 @@ func NewCue(filename string) (*HostConfig, error) {
 	if err := unified.Validate(cue.Concrete(true)); err != nil {
 		return nil, err
 	}
-	conf, err := NewViper()
-	if err != nil {
+
+	if base == nil {
+		base = &ExperimentConfig{}
+	}
+
+	if err := cfg.Decode(base); err != nil {
 		return nil, err
 	}
-	if err := cfg.Decode(conf); err != nil {
-		return nil, err
-	}
-	return conf, nil
+	return base, nil
 }
